# docs/swagger.yaml
basePath: /api/v1
definitions:
  errorHandling.AppError:
    properties:
      code:
        type: integer
      key:
        type: string
      log:
        type: string
      message:
        type: string
    type: object
  models.Address:
    properties:
      Address:
        type: string
      Area:
        $ref: '#/definitions/models.Area'
        type: object
      AreaID:
        type: integer
      ContactName:
        type: string
      Coordinates:
        type: string
      Country:
        type: string
      District:
        type: string
      IsDefault:
        type: boolean
      Phone:
        type: string
      State:
        type: string
      Ward:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Area:
    properties:
      Addresses:
        items:
          $ref: '#/definitions/models.Address'
        type: array
      AreaConfig:
        items:
          $ref: '#/definitions/models.AreaConfig'
        type: array
      Name:
        type: string
      Note:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      updated_at:
        type: integer
    type: object
...
  models.Category:
    properties:
      Image:
        type: string
      Name:
        type: string
      Products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      ProductsStore:
        items:
          $ref: '#/definitions/models.ProductStore'
        type: array
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.DrugStore:
    properties:
      Address:
        $ref: '#/definitions/models.Address'
        type: object
      AddressID:
        $ref: '#/definitions/models.UID'
        type: object
      ApproveBy:
        $ref: '#/definitions/models.User'
        type: object
      ApproveTime:
        type: integer
      ChildStores:
        items:
          $ref: '#/definitions/models.DrugStore'
        type: array
      LicenseFile:
        type: string
      Orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      OrdersStore:
        items:
          $ref: '#/definitions/models.OrderStore'
        type: array
      Phone:
        type: string
      Products:
        items:
          $ref: '#/definitions/models.ProductStore'
        type: array
      Representative:
        $ref: '#/definitions/models.User'
        type: object
      Staff:
        $ref: '#/definitions/models.User'
        type: object
      Status:
        type: string
      StoreName:
        type: string
      TaxNumber:
        type: string
      Type:
        type: string
      Users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      Vouchers:
        items:
          $ref: '#/definitions/models.Voucher'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Partner:
    properties:
      Address:
        $ref: '#/definitions/models.Address'
        type: object
      AddressID:
        $ref: '#/definitions/models.UID'
        type: object
      Email:
        type: string
      Name:
        type: string
      Note:
        type: string
      Representative:
        $ref: '#/definitions/models.User'
        type: object
      Status:
        type: string
      Type:
        type: string
      Users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Product:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      ConditionVoucher:
        type: string
      Content:
        type: string
      Cost:
        type: number
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      HasPromote:
        type: boolean
      HasPromoteVoucher:
        type: boolean
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      Percent:
        type: number
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      ValueVoucher:
        type: number
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      Voucher:
        $ref: '#/definitions/models.Voucher'
        type: object
      VoucherId:
        $ref: '#/definitions/models.UID'
        type: object
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.ProductStore:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      Content:
        type: string
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Role:
    properties:
      Description:
        type: string
      RoleName:
        type: string
      Users:
        items:
          $ref: '#/definitions/models.User'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: integer
    type: object
...
  models.UID:
    type: object
  models.User:
    properties:
      Address:
        $ref: '#/definitions/models.Address'
        type: object
      CaringDrugstore:
        items:
          $ref: '#/definitions/models.DrugStore'
        type: array
      Confirmed:
        type: boolean
      DrugStore:
        $ref: '#/definitions/models.DrugStore'
        type: object
      Email:
        type: string
      IsAdmin:
        type: boolean
      Name:
        type: string
      Partner:
        $ref: '#/definitions/models.Partner'
        type: object
      Roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      Type:
        type: string
      Username:
        type: string
      carts:
        items:
          $ref: '#/definitions/models.Cart'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  requests.AddressRequest:
    properties:
      Address:
        example: Quan Hoa
        type: string
      AreaID:
        $ref: '#/definitions/models.UID'
        type: object
      ContactName:
        example: Jackie
        type: string
      Coordinates:
        example: Jackie
        type: string
      Country:
        example: Vietnam
        type: string
      District:
        example: Cầu giấy
        type: string
      Id:
        $ref: '#/definitions/models.UID'
        type: object
      IsDefault:
        example: true
        type: boolean
      Phone:
        example: "0345532343"
        type: string
      State:
        example: Hà nội
        type: string
      Ward:
        example: Quan Hoa
        type: string
    required:
    - Address
    - AreaID
    - ContactName
    - Coordinates
    - Country
    - District
    - IsDefault
    - Phone
    - State
    - Ward
    type: object
...
  requests.CategoryRequest:
    properties:
      Name:
        example: name
        type: string
      Note:
        example: note
        type: string
      Slug:
        example: slug
        type: string
    type: object
...
  requests.CreatePartnerRequest:
    properties:
      Address:
        $ref: '#/definitions/requests.AddressRequest'
        type: object
      Email:
        example: abc@gmail.com
        type: string
      Name:
        example: MeTri
        type: string
      Note:
        example: acbasd
        type: string
      Status:
        example: new
        type: string
      Type:
        example: supplier/manufacturer
        type: string
    required:
    - Address
    - Email
    - Name
    - Note
    - Status
    - Type
    type: object
...
  requests.EditAddressRequest:
    properties:
      Address:
        example: Quan Hoa
        type: string
      AreaID:
        $ref: '#/definitions/models.UID'
        type: object
      ContactName:
        example: Jackie
        type: string
      Coordinates:
        example: Jackie
        type: string
      Country:
        example: Vietnam
        type: string
      District:
        example: Cầu giấy
        type: string
      Id:
        $ref: '#/definitions/models.UID'
        type: object
      IsDefault:
        example: true
        type: boolean
      Phone:
        example: "0345532343"
        type: string
      State:
        example: Hà nội
        type: string
      Ward:
        example: Quan Hoa
        type: string
    required:
    - Address
    - AreaID
    - ContactName
    - Coordinates
    - Country
    - District
    - Id
    - IsDefault
    - Phone
    - State
    - Ward
    type: object
...
  requests.EditPartnerRequest:
    properties:
      Address:
        $ref: '#/definitions/requests.EditAddressRequest'
        type: object
      Email:
        example: abc@gmail.com
        type: string
      Name:
        example: MeTri
        type: string
      Note:
        example: acbasd
        type: string
      Status:
        example: new
        type: string
      Type:
        example: supplier/manufacturer
        type: string
    required:
    - Address
    - Email
    - Name
    - Note
    - Status
    - Type
    type: object
...
  requests.SearchCategoryRequest:
    properties:
      Name:
        example: name
        type: string
      Slug:
        example: slug
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
    type: object
...
  requests.SearchPartnerRequest:
    properties:
      Name:
        example: MeTri
        type: string
      Status:
        example: active
        type: string
      Type:
        example: parent
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
      time_from:
        type: integer
      time_to:
        type: integer
    type: object
...
  requests.SortOption:
    properties:
      sort_direction:
        type: string
      sort_field:
        type: string
    type: object
...
  responses.CategorySearch:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
...
  responses.Data:
    properties:
      code:
        type: integer
      key:
        type: string
      message:
        type: string
    type: object
...
  responses.PartnerSearch:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Partner'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
...
host: v1.api.medilane.vn
info:
  contact:
    name: medilane team
    url: https://www.medilane.vn/
  description: This is openapi for account api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Medilane account api
  version: "1.0"
paths:
...
  /category:
    post:
      consumes:
      - application/json
      description: Perform create Category
      operationId: create-category
      parameters:
      - description: Filter Category
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create category in system
      tags:
      - Category-Management
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Perform delete Category
      operationId: delete-category
      parameters:
      - description: id Category
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Delete category in system
      tags:
      - Category-Management
    put:
      consumes:
      - application/json
      description: Perform edit Category
      operationId: edit-category
      parameters:
      - description: body Category
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CategoryRequest'
      - description: id Category
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit category in system
      tags:
      - Category-Management
  /category/find:
    post:
      consumes:
      - application/json
      description: Perform search Category
      operationId: search-category
      parameters:
      - description: Filter Category
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CategorySearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search category in system
      tags:
      - Category-Management
...
  /partner:
    post:
      consumes:
      - application/json
      description: Perform create partner
      operationId: create-partner
      parameters:
      - description: Filter partner
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreatePartnerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create partner in system
      tags:
      - Partner Management
  /partner/{id}:
    delete:
      consumes:
      - application/json
      description: Perform partner role
      operationId: delete-partner
      parameters:
      - description: id partner
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Delete partner in system
      tags:
      - Partner Management
    get:
      consumes:
      - application/json
      description: Perform get partner
      operationId: get-partner
      parameters:
      - description: id partner
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Partner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Get partner in system
      tags:
      - Partner Management
    put:
      consumes:
      - application/json
      description: Perform edit partner
      operationId: edit-partner
      parameters:
      - description: body partner
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.EditPartnerRequest'
      - description: id partner
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit partner in system
      tags:
      - Partner Management
  /partner/find:
    post:
      consumes:
      - application/json
      description: Perform search partner
      operationId: search-partner
      parameters:
      - description: search partner
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchPartnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.PartnerSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search partner in system
      tags:
      - Partner Management
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"